backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
google-auth==2.36.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.137.0
pydantic-settings==2.6.1
pypdf==4.2.0
python-docx==1.1.2
openpyxl==3.1.5
pytest==8.3.3
